refactor:#48 fetch 함수 분리
